+++ Source/searchProfile.py
@@ -59,7 +59,7 @@ def removeProfile(name: str):
     connection.commit()
     connection.close()
 
-def addProfile(name: str, options: searchOptions):
+def addProfile(name: str, options: searchOptions.SearchOptions):
     connection = sqlite3.connect(kesDBPath)
     cursor = connection.cursor()
     profileID = _getProfileID(name, cursor)
@@ -67,7 +67,16 @@ def addProfile(name: str, options: searchOptions):
     connection.commit()
     connection.close()
 
-def getProfile(name: str):
+def updateProfile(name: str, options: searchOptions.SearchOptions):
+    connection = sqlite3.connect(kesDBPath)
+    cursor = connection.cursor()
+    profileID = _getProfileID(name, cursor)
+    _deleteAllOptions(profileID, cursor)
+    _addOptions(profileID, options, cursor)
+    connection.commit()
+    connection.close()
+
+def getProfile(name: str) -> searchOptions.SearchOptions:
     connection = sqlite3.connect(kesDBPath)
     cursor = connection.cursor()
     profileID = _getProfileID(name, cursor)
@@ -82,7 +91,7 @@ def getProfile(name: str):
 
 # get the id of the profile based on the name
 # create one if it does not exist
-def _getProfileID(name: str, cursor: sqlite3.Cursor):
+def _getProfileID(name: str, cursor: sqlite3.Cursor) -> int:
     profiles = cursor.execute("SELECT id FROM profile WHERE name = '{}';".format(name)).fetchall()
     if len(profiles) == 0:
         profiles = cursor.execute("INSERT INTO profile (name) VALUES ('{}');".format(name)).fetchall()
@@ -114,6 +123,14 @@ def _addGenres(profileID: int, genres, cursor: sqlite3.Cursor):
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
 
+def _getGenres(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'genre';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
+
 # Trigger CRUD Operations for tags
 def _addTags(profileID: int, tags, cursor: sqlite3.Cursor):
     if len(tags) != 0:
@@ -122,6 +139,14 @@ def _addTags(profileID: int, tags, cursor: sqlite3.Cursor):
         for x in tags:
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
+
+def _getTags(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'tag';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
 
 # Trigger CRUD Operations for cast
 def _addCast(profileID: int, cast, cursor: sqlite3.Cursor):
@@ -132,6 +157,14 @@ def _addCast(profileID: int, cast, cursor: sqlite3.Cursor):
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
 
+def _getCast(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'cast';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
+
 # Trigger CRUD Operations for director
 def _addDirector(profileID: int, director, cursor: sqlite3.Cursor):
     if len(director) != 0:
@@ -140,6 +173,14 @@ def _addDirector(profileID: int, director, cursor: sqlite3.Cursor):
         for x in director:
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
+
+def _getDirector(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'director';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
 
 # Trigger CRUD Operations for include list
 def _addIncludes(profileID: int, includes, cursor: sqlite3.Cursor):
@@ -150,6 +191,14 @@ def _addIncludes(profileID: int, includes, cursor: sqlite3.Cursor):
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
 
+def _getIncludes(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'include';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
+
 # Trigger CRUD Operations for exclude list
 def _addExcludes(profileID: int, excludes, cursor: sqlite3.Cursor):
     if len(excludes) != 0:
@@ -159,8 +208,16 @@ def _addExcludes(profileID: int, excludes, cursor: sqlite3.Cursor):
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
 
+def _getExcludes(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'exclude';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
+
 # Trigger CRUD Operations for Watch Status
-def _addWatchStatus(profileID: int, watchStatus, cursor: sqlite3.Cursor):
+def _addWatchStatus(profileID: int, watchStatus: int, cursor: sqlite3.Cursor):
     if len(watchStatus) != 0:
         valueList = ""
         defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'watchStatus';").fetchall()[0][0]
@@ -168,11 +225,23 @@ def _addWatchStatus(profileID: int, watchStatus, cursor: sqlite3.Cursor):
             valueList = "({},{},{})".format(x,profileID,defaultID) if valueList == "" else "{},({},{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
 
+def _getWatchStatus(profileID:int, cursor: sqlite3.Cursor) -> int:
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'watchStatus';").fetchall()[0][0]
+    return cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()[0][0]
+
 # Trigger CRUD Operations for length
 def _addLength(profileID: int, length, cursor: sqlite3.Cursor):
         minID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'lengthMIN';").fetchall()[0][0]
         maxID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'lengthMAX';").fetchall()[0][0]
         cursor.execute("INSERT INTO profileOptions (value, r_profile, r_searchOptionType) VALUES ({},{},{}), ({},{},{});".format(length[0],profileID,minID,length[1],profileID,maxID))
+
+def _getMinLength(profileID:int, cursor: sqlite3.Cursor) -> int:
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'lengthMIN';").fetchall()[0][0]
+    return cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()[0][0]
+
+def _getMaxLength(profileID:int, cursor: sqlite3.Cursor) -> int:
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'lengthMAX';").fetchall()[0][0]
+    return cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()[0][0]
 
 # Trigger CRUD Operations for year
 def _addYears(profileID: int, years, cursor: sqlite3.Cursor):
@@ -183,6 +252,14 @@ def _addYears(profileID: int, years, cursor: sqlite3.Cursor):
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
 
+def _getYears(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'year';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
+
 # Trigger CRUD Operations for rating
 def _addRating(profileID: int, rating, cursor: sqlite3.Cursor):
     if len(rating) != 0:
@@ -191,6 +268,14 @@ def _addRating(profileID: int, rating, cursor: sqlite3.Cursor):
         for x in rating:
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
+
+def _getRating(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'rating';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
 
 # Trigger CRUD Operations for media type
 def _addMediaType(profileID: int, mediaType, cursor: sqlite3.Cursor):
@@ -201,6 +286,14 @@ def _addMediaType(profileID: int, mediaType, cursor: sqlite3.Cursor):
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
 
+def _getMediaType(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'mediaType';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
+
 # Trigger CRUD Operations for studios
 def _addStudios(profileID: int, studios, cursor: sqlite3.Cursor):
     if len(studios) != 0:
@@ -210,22 +303,49 @@ def _addStudios(profileID: int, studios, cursor: sqlite3.Cursor):
             valueList = "('{}',{},{})".format(x,profileID,defaultID) if valueList == "" else "{},('{}',{},{})".format(valueList, x,profileID,defaultID)
         cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES {};".format(valueList))
 
+def _getStudios(profileID:int, cursor: sqlite3.Cursor):
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'studio';").fetchall()[0][0]
+    tempList = cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()
+    returnList = []
+    for x in tempList:
+        returnList.append(x[0])
+    return returnList
+
 # Trigger CRUD Operations for Most Watched
 def _addMostWatched(profileID: int, mostWatched, cursor: sqlite3.Cursor):
     defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'mostWatched';").fetchall()[0][0]
     cursor.execute("INSERT INTO profileOptions (value,r_profile,r_searchOptionType) VALUES ('{}',{},{});".format(mostWatched,profileID,defaultID))
 
+def _getMostWatched(profileID:int, cursor: sqlite3.Cursor) -> bool:
+    defaultID = cursor.execute("SELECT id FROM searchOptionType WHERE title = 'mostWatched';").fetchall()[0][0]
+    return cursor.execute("SELECT value FROM profileOptions WHERE r_profile = {} AND r_searchOptionType = {};".format(profileID, defaultID)).fetchall()[0][0]
+
+# Delete Functions
 def _deleteAllOptions(profileID: int, cursor: sqlite3.Cursor):
     cursor.execute("DELETE FROM profileOptions WHERE r_profile = {}".format(profileID))
 
 def _deleteProfile(profileID: int, cursor: sqlite3.Cursor):
     cursor.execute("DELETE FROM profile WHERE id = {}".format(profileID))
 
-def _getProfileOptions(profileID: int, cursor: sqlite3.Cursor):
-    profile = searchOptions()
-
+def _getProfileOptions(profileID: int, cursor: sqlite3.Cursor) -> searchOptions.SearchOptions:
+    profile = searchOptions.SearchOptions()
+    
     # this will call a bunch of _get* functions to get the array of values for each property
-
+    profile.setGenre(_getGenres(profileID, cursor))
+    profile.setTag(_getTags(profileID, cursor))
+    profile.setCast(_getCast(profileID, cursor))
+    profile.setDirector(_getDirector(profileID, cursor))
+    profile.setInclude(_getIncludes(profileID, cursor))
+    profile.setExclude(_getExcludes(profileID, cursor))
+    profile.setWatchStatus(_getWatchStatus(profileID, cursor))
+    profile.setMINLength(_getMinLength(profileID, cursor))
+    profile.setMAXLength(_getMaxLength(profileID, cursor))
+    profile.setYear(_getYears(profileID, cursor))
+    profile.setRating(_getRating(profileID, cursor))
+    profile.setMediaType(_getMediaType(profileID, cursor))
+    profile.setStudio(_getStudios(profileID, cursor))
+    profile.setMostWatched(_getMostWatched(profileID, cursor))
+    
     return profile
 
 # test cases
@@ -233,6 +353,9 @@ def unitTest(options: searchOptions):
 
     addProfile('dustin', options)
     print('Profile Added')
+
+    testProfile = getProfile('dustin')
+    print("Get Profile: {}".format(testProfile))
 
     removeProfile('dustin')
     print('Profile Deleted')